+++ 09-Stuyablo/Ogre.java
@@ -1,5 +1,5 @@
 public class Ogre extends BaseChar{
-    private int healthBoost = 15;
+    private double healthBoost = 15;
     public void boost(){
         super.addmaxhealth(this.healthBoost);
     }
